# lms-openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/courses/{id}/publishing/submit": {
            "post": {
                "description": "Author submits a DRAFT or CHANGES_REQUESTED course for approval. Moves status to IN_REVIEW.",
                "operationId": "CoursePublishingController_submitForReview",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Course product ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitForReviewDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course submitted for review"
                    },
                    "400": {
                        "description": "Invalid transition or missing submission"
                    },
                    "403": {
                        "description": "Not the course author"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Submit course for manager review",
                "tags": [
                    "Course Publishing"
                ]
            }
        },
        "/courses/{id}/publishing/approve": {
            "post": {
                "description": "Moves status from IN_REVIEW → APPROVED.",
                "operationId": "CoursePublishingController_approve",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Course product ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewDecisionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course approved"
                    },
                    "400": {
                        "description": "Invalid transition"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Approve a submitted course",
                "tags": [
                    "Course Publishing"
                ]
            }
        },
        "/courses/{id}/publishing/request-changes": {
            "post": {
                "description": "Moves status from IN_REVIEW → CHANGES_REQUESTED. A review comment is required.",
                "operationId": "CoursePublishingController_requestChanges",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Course product ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewDecisionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Changes requested"
                    },
                    "400": {
                        "description": "Comment required or invalid transition"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Request changes on a submitted course",
                "tags": [
                    "Course Publishing"
                ]
            }
        },
        "/courses/{id}/publishing/publish": {
            "post": {
                "description": "Moves status from APPROVED → ON_AIR.",
                "operationId": "CoursePublishingController_publish",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Course product ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Course published and now ON_AIR"
                    },
                    "400": {
                        "description": "Course not in APPROVED state"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Publish an approved course",
                "tags": [
                    "Course Publishing"
                ]
            }
        },
        "/courses/{id}/publishing/unpublish": {
            "post": {
                "description": "Moves status from ON_AIR → DRAFT.",
                "operationId": "CoursePublishingController_unpublish",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Course product ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Course unpublished"
                    },
                    "400": {
                        "description": "Course not ON_AIR"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Unpublish a live course",
                "tags": [
                    "Course Publishing"
                ]
            }
        },
        "/courses/{id}/publishing/archive": {
            "post": {
                "description": "Moves course to ARCHIVED (terminal state).",
                "operationId": "CoursePublishingController_archive",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Course product ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Course archived"
                    },
                    "400": {
                        "description": "Invalid state for archiving"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Archive a course",
                "tags": [
                    "Course Publishing"
                ]
            }
        },
        "/courses/{id}/publishing/submissions": {
            "get": {
                "description": "Returns all past submission cycles for the course.",
                "operationId": "CoursePublishingController_getSubmissionHistory",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Course product ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission history"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get submission history for a course",
                "tags": [
                    "Course Publishing"
                ]
            }
        },
        "/course-publishing/review-queue": {
            "get": {
                "description": "Returns courses pending review (or filtered by submission status).",
                "operationId": "CoursePublishingQueueController_getReviewQueue",
                "parameters": [
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "description": "Filter by submission status",
                        "schema": {
                            "default": "PENDING",
                            "type": "string",
                            "enum": [
                                "PENDING",
                                "APPROVED",
                                "CHANGES_REQUESTED"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "default": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "default": 20,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated review queue"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get the manager review queue",
                "tags": [
                    "Course Publishing"
                ]
            }
        },
...
    },
    "info": {
        "title": "VirtuaLMS API",
        "description": "The VirtuaLMS API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "SubmitForReviewDto": {
                "type": "object",
                "properties": {
                    "note": {
                        "type": "string",
                        "description": "Optional note from the author when submitting",
                        "example": "All sections and videos are complete. Ready for review."
                    }
                }
            },
            "ReviewDecisionDto": {
                "type": "object",
                "properties": {
                    "comment": {
                        "type": "string",
                        "description": "Reviewer comment — required when requesting changes, optional when approving",
                        "example": "Great course! Please fix the typo in Section 2 session title."
                    }
                }
            },
...
        }
    }
}